Move selectors to component
